ADD OutNode / Switch : node witch are not switchable by home automation It allow reduce margin because we can count heavy consumer that are not switchable to avoid black-out (on off-peak time-slots).
