+++ config/openhems.yaml
@@ -20,5 +20,5 @@ network:
         contract: {class: RTETempo}
       }
     - {id: voiture, strategy: offpeak, class: switch, isOn: switch.tz3000_2putqrmw_ts011f_commutateur_2, currentPower: sensor.tz3000_2putqrmw_ts011f_puissance_2, maxPower: 2300}
-    - {id: box, strategy: redDay, class: switch, isOn: switch.tz3000_2putqrmw_ts011f_commutateur, currentPower: sensor.tz3000_2putqrmw_ts011f_puissance}
+    - {id: box, strategy: redDay, class: out, currentPower: sensor.tz3000_2putqrmw_ts011f_puissance}
