+++ src/openhems/modules/network/__init__.py
@@ -6,4 +6,9 @@ It is used to know devices and to switch on/off them.
 from .homestate_updater import HomeStateUpdater, HomeStateUpdaterException
 from .network import OpenHEMSNetwork
 from .node import OpenHEMSNode
-from .feeder import Feeder, SourceFeeder, ConstFeeder
+from .outnode import OutNode, Switch
+from .inoutnode import InOutNode, PublicPowerGrid, SolarPanel, Battery
+from .feeder import (
+	Feeder, SourceFeeder, RandomFeeder, ConstFeeder,
+	RotationFeeder, FakeSwitchFeeder, StateFeeder, SumFeeder # For fakeNetwork
+)
