+++ src/openhems/modules/network/node.py
@@ -6,10 +6,8 @@ import logging
 
 from typing import Final
 from collections import deque, OrderedDict
-from openhems.modules.web import OpenHEMSSchedule
-from openhems.modules.contract import Contract
 from openhems.modules.util import CastUtililty, ConfigurationException
-from .feeder import Feeder, ConstFeeder
+from .feeder import Feeder
 
 CYCLE_HISTORY: Final[int] = 10 # Number of cycle we keep history
 logger = logging.getLogger(__name__)
@@ -23,6 +21,7 @@ class OpenHEMSNode:
 		"""
 		Set Home-Assistant id
 		"""
+		self.name = haId
 		self.id = haId.strip().replace(" ", "_")
 
 	def __init__(self, nameId, currentPower, maxPower, isOnFeeder=None,
@@ -238,233 +237,3 @@ class OpenHEMSNode:
 			return ok
 		logger.warning("Try to switchOn/Off a not switchable device : %s", self.id)
 		return connect # Consider node is always on network
-
-class OutNode(OpenHEMSNode):
-	"""
-	Electricity consumer (like washing-machine, water-heater).
-	:param int priority: A device with higth priority is more important than a low priority one.
-		Usually priority is a number between 0 and 100
-	"""
-	def __init__(self, nameId, strategyId, currentPower, maxPower, isOnFeeder=None,
-			priority=50, network=None):
-		super().__init__(nameId, currentPower, maxPower, isOnFeeder, network=network)
-		self.name = nameId
-		self.schedule = OpenHEMSSchedule(self.id, nameId, self)
-		self.strategyId = strategyId
-		self._priority = priority
-
-	def getPriority(self):
-		"""
-		:return int: number representing the level of priority
-		"""
-		return self._priority
-
-	def getSchedule(self):
-		"""
-		Return schedule
-		"""
-		return self.schedule
-
-	def setCondition(self, condition):
-		"""
-		Set a condition to switchOn
-		even if the node is  not manually schedule.
-		"""
-		self.schedule.setCondition(condition)
-		return condition
-
-	def isScheduled(self):
-		"""
-		Return True, if device is schedule to be on
-		"""
-		sch = self.getSchedule()
-		if sch is not None:
-			return sch.isScheduled()
-		return False
-
-	def decrementTime(self, time:int) -> int:
-		"""
-		Decrease time of schedule and return remaining time.
-		"""
-		sch = self.getSchedule()
-		if sch is not None and self.isOn():
-			return sch.decrementTime(time)
-		return 0
-
-	def getStrategyId(self):
-		"""
-		Return StrategyId
-		"""
-		return self.strategyId
-
-	def __str__(self):
-		return (f"OutNode(name={self.name}, strategy={self.strategyId}, priority={self._priority}"
-			f" currentPower={self.currentPower},"
-			f"maxPower={self.maxPower}, isOn={self.isOn()})")
-	def __repr__(self):
-		return str(self)
-
-class InOutNode(OpenHEMSNode):
-	"""
-	It is electricity source, it may consume electricity over-production
-	 if possible (Battery with MPPT or Sell on public-grid)
-	param maxPower: positive value, max power we can consume at a time.
-	param minPower: negative value if we can sell or ther is battery, 0 overwise.
-	"""
-	def __init__(self, nameid, currentPower, maxPower, minPower, marginPower) -> None:
-		# isAutoAdatative: bool, isControlable: bool, isModulable: bool, isCyclic: bool
-		super().__init__(nameid, currentPower, maxPower)
-		self.currentPower = currentPower
-		self.marginPower = marginPower
-		self.minPower = minPower
-
-	def respectConstraints(self, power=None):
-		"""
-		Check min/max constraints for power
-		
-		return bool: true if 'power' respects constraints
-		"""
-		if power is None:
-			power = self.currentPower.getValue()
-
-		if power+self.marginPower.getValue()>self.maxPower.getValue():
-			return False
-		if power-self.marginPower.getValue()<self.minPower.getValue():
-			return False
-		return True
-
-	def getMinPower(self):
-		"""
-		Return current minimal power
-		"""
-		return self.minPower.getValue()
-	def getMarginPower(self):
-		"""
-		Return current margin power
-		"""
-		margin = self.marginPower.getValue()
-		# logger.debug("MarginPower of Node %s is %s", self.id, margin)
-		return margin
-
-	def _getSafetyLevel(self):
-		"""
-		Get a int value representing how safe is the current power value
-
-		return int:
-			- 0: unsafe
-			- 1: respect constraints but shouldn't on next loop
-			- 2: respect constraints but could be out of constraints next loop
-			- 3: Safe values
-		"""
-		if not self.respectConstraints():
-			return 0
-		_min, avg, _max = self._estimateNextPower()
-		if not self.respectConstraints(avg):
-			return 1
-		if not (self.respectConstraints(_min) or self.respectConstraints(_max)):
-			return 2
-		return 3
-
-class PublicPowerGrid(InOutNode):
-	"""
-	This represent Public power grid. Just one should be possible.
-	"""
-	def __init__(self, nameid, currentPower, maxPower, minPower, marginPower,
-	             contract, networkUpdater):
-		super().__init__(nameid, currentPower, maxPower, minPower, marginPower)
-		self.contract = Contract.getContract(contract, networkUpdater.conf, networkUpdater)
-
-	def __str__(self):
-		return (f"PublicPowerGrid({self.currentPower}, maxPower={self.maxPower},"
-			f" minPower={self.minPower}, marginPower={self.marginPower}, contract={self.contract})")
-
-	def getContract(self):
-		"""
-		Return the contract. Usefull to get specificities witch can imply on strategy.
-		Like offpeak-hours, prices.
-		"""
-		return self.contract
-
-class SolarPanel(InOutNode):
-	"""
-	This represent photovoltaïc solar panels. 
-	We can have many, but one can represent many solar panel.
-	It depends of sensors number.
-	"""
-	# pylint: disable=too-many-arguments
-	def __init__(self, nameid, currentPower, maxPower, *,
-			moduleModel=None, inverterModel=None, tilt=45, azimuth=180,
-			modulesPerString=1, stringsPerInverter=1, marginPower=None):
-		if marginPower is None:
-			marginPower = ConstFeeder(0)
-		super().__init__(nameid, currentPower, maxPower, currentPower,  marginPower)
-		self.moduleModel = moduleModel
-		self.inverterModel = inverterModel
-		self.tilt = tilt
-		self.azimuth = azimuth
-		self.modulesPerString = modulesPerString
-		self.stringsPerInverter = stringsPerInverter
-
-	def getMaxPower(self):
-		"""
-		get current maximum power = current power.
-		value saved in maxPower, is the theorical maxPower (usefull to know efficiency).
-		But in fact even if we ask more power, solar panels can't give us more.
-		"""
-		return self.currentPower.getValue()
-
-	def __str__(self):
-		return (f"SolarPanel({self.currentPower}, {self.maxPower},"
-		f" moduleModel={self.moduleModel}, inverterModel={self.inverterModel},"
-		f" tilt={self.tilt}, azimuth={self.azimuth},"
-		f" modulesPerString={self.modulesPerString},"
-		f"stringsPerInverter={self.stringsPerInverter})")
-	def __repr__(self):
-		return str(self)
-
-class Battery(InOutNode):
-	"""
-	This represent battery.
-	"""
-	# pylint: disable=too-many-arguments
-	def __init__(self, nameid, capacity, currentPower, *, maxPowerIn=None,
-			maxPowerOut=None, efficiencyIn:float=0.95, efficiencyOut:float=0.95,
-			targetLevel:float=0.70,
-			currentLevel=None, lowLevel:float=0.20, highLevel:float=0.80):
-		if maxPowerIn is None:
-			maxPowerIn = ConstFeeder(2300) # a standard electrical outlet
-		if maxPowerOut is None:
-			maxPowerOut = ConstFeeder(-1*maxPowerIn.getValue())
-		super().__init__(nameid, currentPower, maxPowerIn, maxPowerOut, 0)
-		self.isControlable = True
-		self.isModulable = False
-		self.capacity = capacity
-		self.currentLevel = currentLevel
-		self.lowLevel = lowLevel
-		self.highLevel = highLevel
-		self.targetLevel = targetLevel
-		self.efficiencyIn = efficiencyIn
-		self.efficiencyOut = efficiencyOut
-
-	def getCapacity(self):
-		"""
-		Get battery max capacity.
-		"""
-		return self.capacity.getValue()
-
-	def getLevel(self):
-		"""
-		Get battery level.
-		"""
-		return self.currentLevel.getValue()
-
-	def __str__(self):
-		return (f"Battery(capacity={self.capacity}, currentPower={self.currentPower},"
-			f" maxPowerIn={self.maxPower}, maxPowerOut={self.minPower},"
-			f" efficiencyIn={self.efficiencyIn}, level={self.currentLevel},"
-			f" lowLevel={self.lowLevel}, highLevel={self.highLevel})")
-
-	def __repr__(self):
-		return str(self)
-# class CarCharger(Switch):
-# class WaterHeater(InOutNode):
