+++ src/openhems/modules/network/outnode.py
@@ -1,0 +1,139 @@
+"""
+Represent device of home network
+"""
+
+import logging
+
+from openhems.modules.web import OpenHEMSSchedule
+from openhems.modules.util import ConfigurationException
+from .feeder import Feeder
+from .node import OpenHEMSNode
+
+class GuessIsOnFeeder(Feeder):
+	"""
+	Guess if node is on with currentPower.
+	"""
+	def __init__(self, source, nbCycleWithoutPowerForOff=1):
+		"""
+		:param source: The node source of power with network.getCycleId() and getCurrentPower().
+		:param nbCycleWithoutPowerForOff: number of cycle without power before we consider the device is off
+		"""
+		super().__init__(source)
+		self._value = source
+		self._nbCycleWithoutPower = 0
+		self._lastCycleId = -1
+		# Number of cycle without power before we consider the device is off
+		self.nbCycleWithoutPowerForOff = nbCycleWithoutPowerForOff
+
+	def getValue(self):
+		"""
+		Return True if the device is on (consuming power).
+		"""
+		if self._value.getCurrentPower()>0:
+			self._nbCycleWithoutPower = 0
+			return True
+		cycleId = self._value.network.getCycleId()
+		if self._lastCycleId!=cycleId:
+			self._nbCycleWithoutPower += 1
+			self._lastCycleId=cycleId
+		return self._nbCycleWithoutPower<self.nbCycleWithoutPowerForOff
+
+	def __str__(self):
+		return f"GuessIsOnFeeder(isOn={self.getValue()}, nbCycleWithoutPowerForOff={self.nbCycleWithoutPowerForOff})"
+
+	def __repr__(self):
+		return str(self)
+
+class OutNode(OpenHEMSNode):
+	"""
+	Electricity consumer (like washing-machine, water-heater) that we can't switch on/off.
+	We can just know if it has a power consumption. We deduce from it if it's on or off.
+	This is usefull for appliances not switchable by home automation but wich can consum a lot.
+	This is to take care of their max consumption in OpenHEMSServer.check() to avoid over-load.
+	"""
+	def __init__(self, nameId, currentPower, maxPower, network=None, nbCycleWithoutPowerForOff=0):
+		isOn = GuessIsOnFeeder(self, nbCycleWithoutPowerForOff)
+		super().__init__(nameId, currentPower, maxPower, network=network, isOnFeeder=isOn)
+
+	def isSwitchable(self):
+		return False
+
+	def __str__(self):
+		return (f"OutNode(name={self.name}, currentPower={self.currentPower},"
+			f"maxPower={self.maxPower}, isOn={self.isOn()})")
+
+	def __repr__(self):
+		return str(self)
+
+class Switch(OutNode):
+	"""
+	Electricity consumer (like washing-machine, water-heater)
+	wich can be switch on/off.
+	"""
+	def __init__(self, nameId, strategyId, currentPower, maxPower, isOnFeeder=None,
+			priority=50, network=None):
+		if isOnFeeder is None:
+			raise ConfigurationException("Declare a Switch() but without seting isOn.")
+		super(OutNode, self).__init__(nameId, currentPower, maxPower, isOnFeeder=isOnFeeder, network=network)
+		self.schedule = OpenHEMSSchedule(self.id, nameId, self)
+		self.strategyId = strategyId
+		self._priority = priority
+
+	def isSwitchable(self):
+		return True
+
+	def isOn(self):
+		"""
+		Return True if the device is on (consuming power).
+		"""
+		# useless, but as GuessIsOnFeeder could be replace by overiding isOn() method
+		return super(OutNode, self).isOn()
+
+	def getPriority(self):
+		"""
+		:return int: number representing the level of priority
+		"""
+		return self._priority
+
+	def getSchedule(self):
+		"""
+		Return schedule
+		"""
+		return self.schedule
+
+	def setCondition(self, condition):
+		"""
+		Set a condition to switchOn
+		even if the node is  not manually schedule.
+		"""
+		self.schedule.setCondition(condition)
+		return condition
+
+	def isScheduled(self):
+		"""
+		Return True, if device is schedule to be on
+		"""
+		sch = self.getSchedule()
+		if sch is not None:
+			return sch.isScheduled()
+		return False
+
+	def decrementTime(self, time:int) -> int:
+		"""
+		Decrease time of schedule and return remaining time.
+		"""
+		sch = self.getSchedule()
+		if sch is not None and self.isOn():
+			return sch.decrementTime(time)
+		return 0
+
+	def getStrategyId(self):
+		"""
+		Return StrategyId
+		"""
+		return self.strategyId
+
+	def __str__(self):
+		return (f"Switch(name={self.name}, strategy={self.strategyId}, priority={self._priority}"
+			f" currentPower={self.currentPower},"
+			f"maxPower={self.maxPower}, isOn={self.isOn()})")
